refactor: Move methods to the helpers file and update getFreePort definition
